update configs, merge model into single file
