+++ packages/@n8n/ai-node-sdk/tsconfig.json
@@ -1,5 +1,8 @@
 {
-	"extends": "@n8n/typescript-config/tsconfig.common.json",
+	"extends": [
+		"@n8n/typescript-config/tsconfig.common.json",
+		"@n8n/typescript-config/tsconfig.backend.json"
+	],
 	"compilerOptions": {
 		"rootDir": ".",
 		"baseUrl": ".",
